Add soft delete learning material & learning material type

+++ elearning/src/main/java/com/lawencon/elearning/service/LearningMaterialsServiceImpl.java
@@ -1,6 +1,7 @@
 package com.lawencon.elearning.service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -74,8 +75,19 @@ public class LearningMaterialsServiceImpl extends BaseServiceImpl implements Lea
 	}
 
 	@Override
-	public void deleteLearningMaterialById(String id) throws Exception {
-		learningMaterialsDao.deleteLearningMaterialById(id);
+	public void deleteLearningMaterialById(String id, String idUser) throws Exception {
+		try {
+			begin();
+			if(validateDelete(id)) {
+				learningMaterialsDao.softDeleteLearningMaterialById(id, idUser);
+			} else {
+				learningMaterialsDao.deleteLearningMaterialById(id);				
+			}
+			commit();
+		} catch(Exception e) {
+			e.getMessage();
+			rollback();
+		}
 	}
 
 	@Override
@@ -171,5 +183,14 @@ public class LearningMaterialsServiceImpl extends BaseServiceImpl implements Lea
 				}
 			}
 		}
+	}
+	
+	private boolean validateDelete(String id) throws Exception {
+		List<?> listObj = learningMaterialsDao.validateDeleteLearningMaterial(id);
+		listObj.forEach(System.out::println);
+		List<?> list =  listObj.stream().filter(val -> val != null)
+				.collect(Collectors.toList());
+		System.out.println(list.size());
+		return list.size() > 0 ? true : false;
 	}
 }
